Adjust integrate output for ghc-9
ghc-9 introduces a regression(?) in FP precision (See ghc ticket #19569).
While this is unfortunate we want nofib to be useable with ghc 8.10 and
9.0 as far as possible.

So I added an additional argument to the benchmark which controls if we
output the benchmark result or not, which is off by default.

This allows easy checking of the output manually without causing
benchmark errors because the output differs.

I've also added the benchmark to the ghc testsuite to avoid future
regressions.